Moved from float to uint32_t for TX power, substituted dBm with mW in CPS channel struct
General code optimization allowing to reduce the overall binary size, especially
for devices without hardware floating point support. On the MD-380 (which has the
FPU) the .text size has been reduced of 4'928 bytes.

+++ platform/drivers/baseband/SA8x8.h
@@ -58,9 +58,9 @@ int sa8x8_enableHSMode();
 /**
  * Set the transmission power.
  *
- * @param power: transmission power in Watt.
+ * @param power: transmission power in mW.
  */
-void sa8x8_setTxPower(const float power);
+void sa8x8_setTxPower(const uint32_t power);
 
 /**
  * Enable or disable the speaker power amplifier.
